requirements.txt
flask
waitress
openai
python-dotenv
